001 feat 주문 시스템 구현

+++ src/main/java/com/mysite/cuffee/order/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.mysite.cuffee.order.repository;
 import com.mysite.cuffee.order.entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderRepository extends JpaRepository<OrderItem, Long> {
 }
